+++ store.go
@@ -23,6 +23,12 @@ type Store interface {
 	DeleteChampionship(id string) error
 
 	// Live Timings
+	UpsertLiveTimingsData(*LiveTimingsPersistedData) error
+	LoadLiveTimingsData() (*LiveTimingsPersistedData, error)
+	UpsertLastRaceEvent(r RaceEvent) error
+	LoadLastRaceEvent() (RaceEvent, error)
+	ClearLastRaceEvent() error
+
 	UpsertLiveFrames([]string) error
 	ListPrevFrames() ([]string, error)