+++ CHANGELOG.md
@@ -3,8 +3,11 @@ v1.5.3
 
 Added:
 
+* Live Timings "Stored Times" will now persist between reboots of Server Manager, if the next event started is the same as the last event running before the Server Manager reboot.
+* You can now configure Server Manager to restart any Event that was running before Server Manager was stopped. Check out "Restart Event On Server Manager Launch" in Server Options.
 * Added an option to prevent web crawlers from accessing any pages on the manager using a robots.txt.
 * Added information about whether a car/track is default/mod/DLC to list and detail pages.
+* Championship and Race Weekend Looping Practice Sessions are now labelled in the navigation bar, e.g. "Championship Practice in Progress"
 * Added penalty options to Qualifying and Practice sessions, penalties can be applied independently in each session. In Race Weekends these penalties will affect the entry list of the next session.
 * Added a button to blacklist a driver directly from results pages, the button is inside the penalties popup and can only be accessed by admins.